+++ test/string_slice_unittest.cpp
@@ -53,11 +53,13 @@ TEST(StringSlice, split)
 {
     auto s = StringSlice("You should name a 分隔符 variable with the same care 分隔符 as you are "
                          "naming your first born child.");
-    auto pieces = s.split("分隔符 ");
-    ASSERT_EQ(pieces.size(), 3);
-    EXPECT_TRUE(pieces[0].equals("You should name a "));
-    EXPECT_TRUE(pieces[1].equals("variable with the same care "));
-    EXPECT_TRUE(pieces[2].equals("as you are naming your first born child."));
+    {
+        auto pieces = s.split("分隔符 ");
+        ASSERT_EQ(pieces.size(), 3);
+        EXPECT_TRUE(pieces[0].equals("You should name a "));
+        EXPECT_TRUE(pieces[1].equals("variable with the same care "));
+        EXPECT_TRUE(pieces[2].equals("as you are naming your first born child."));
+    }
 
     {
         StringSlice pieces[3];