+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
-- 2021-07-02 Add Quaternion.
+- 2021-07-02 Add `nc_abs()`, `nc_min()`, `nc_max()`, `nc_clamp()`
+- 2021-07-02 Add `Quaternion` and `QuaterionD`.
 - 2021-06-30 Add `nc_isnan()`.
 - 2021-06-30 Add `nc_radianToDegree()` and `nc_degreeToRadian()`.
 - 2021-06-30 Add single & double precision PI. NC_PI, NC_PI_2, NC_PI_D, NC_PI_2_D etc.